requirements.txt
#all these modules need to be intalled for covid-stats package to work
#run "pip install -r requirements" to install all these 3 in single run
covid-stats
beautifulsoup4
requests
colorama
python-dotenv


## end of covid-stats package required modules